sdi_geocoder/requirements.txt
requests==2.32.3
python_magic==0.4.27
django-bootstrap4==24.3
django-bootstrap-datepicker-plus==5.0.5